Fix integer type casting in XInsert method for new size calculation

+++ include/XClassArray.h
@@ -592,7 +592,7 @@ protected:
     void XInsert(T *i, T &&o) {
         // Test For Reallocation
         if (m_End == m_AllocatedEnd) {
-            int newsize = (m_AllocatedEnd - m_Begin) * 2;
+            int newsize = (int)((m_AllocatedEnd - m_Begin) * 2);
             if (!newsize)
                 newsize = 1;
             T *newdata = Allocate(newsize);
